pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "image_watermarker"
version = "0.1.0"
description = "A short description of your package"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [ "Pillow", "numpy", "piexif",]
[[project.authors]]
name = "Muhammad Ateeq"
email = "mr.ateeqazam@gmail.com"

["project.urls"]
Homepage = "https://github.com/yourusername/image_watermarker"
Documentation = "https://image_watermarker.readthedocs.io/"

["tool.setuptools.packages.find"]
where = [ ".",]

[project.license]
file = "LICENSE"
